Added crud protection

+++ src/controller/car.js
@@ -15,6 +15,10 @@ export const getCarById = async (req, res) => {
     return res.status(404).json({ message: `No car with id: ${id}` });
   }
 
+  if (car?.userId !== req.body.userId) {
+    return res.status(401).json({ message: `${id} does not belong to you` });
+  }
+
   return res.json({ car: car });
 };
 
@@ -32,26 +36,39 @@ export const insertCar = async (req, res) => {
 export const updateCarById = async (req, res) => {
   const id = req.params.id;
 
+  const existingCar = await CarModel.find({ userId: req.body.userId });
+
+  if (existingCar?.userId !== req.body.userId) {
+    return res.status(401).json({ message: `${id} does not belong to you` });
+  }
+
+  if (!existingCar) {
+    return res.status(404).json({ message: `No car with id: ${id}` });
+  }
+
   const car = await CarModel.findOneAndUpdate(
     { id: id },
     { ...req.body },
     { new: true },
   );
 
-  if (!car) {
-    return res.status(404).json({ message: `No car with id: ${id}` });
-  }
-
   return res.status(200).json({ car: car });
 };
 
 export const deleteCarById = async (req, res) => {
   const id = req.params.id;
-  const car = await CarModel.findOneAndDelete({ id: id });
 
-  if (!car) {
+  const existingCar = await CarModel.find({ userId: req.body.userId });
+
+  if (existingCar?.userId !== req.body.userId) {
+    return res.status(401).json({ message: `${id} does not belong to you` });
+  }
+
+  if (!existingCar) {
     return res.status(404).json({ message: `No car with id: ${id}` });
   }
+
+  const car = await CarModel.findOneAndDelete({ id: id });
 
   return res.status(200).json({ car: car });
 };
